change PanelView to typescript
