+++ app/frontend/src/classes/PanelView/PanelViewFilterAlternativeAlleleFrequency.ts
@@ -1,4 +1,4 @@
-import PanelView from './PanelView.js';
+import PanelView from './PanelView';
 import { storeManager } from '../../store/StoreManager';
 import {
   setSimpleSearchCondition,
