+++ app/frontend/src/classes/PanelView/PanelViewPreviewClinicalSignificance.js
@@ -1,4 +1,4 @@
-import PanelView from './PanelView.js';
+import PanelView from './PanelView.ts';
 import { storeManager } from '../../store/StoreManager';
 import { getSimpleSearchConditionMaster } from '../../store/searchManager';
